backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.32.0.post1
tqdm==4.66.1
tweepy==4.14.0
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
yarl==1.9.4